plug in the final puzzle pieces
